+++ Library/src/main/java/dev/testify/core/processor/compare/FuzzyCompare.kt
@@ -1,7 +1,7 @@
 /*
  * The MIT License (MIT)
  *
- * Modified work copyright (c) 2022 ndtp
+ * Modified work copyright (c) 2022-2024 ndtp
  * Original work copyright (c) 2019 Shopify Inc.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -25,9 +25,11 @@
 package dev.testify.core.processor.compare
 
 import android.graphics.Bitmap
+import android.graphics.Rect
 import com.github.ajalt.colormath.RGB
 import dev.testify.core.TestifyConfiguration
 import dev.testify.core.processor.ParallelPixelProcessor
+import dev.testify.core.processor.ParallelProcessorConfiguration
 import dev.testify.core.processor.compare.colorspace.calculateDeltaE
 
 /**
@@ -38,9 +40,34 @@ import dev.testify.core.processor.compare.colorspace.calculateDeltaE
  *
  * [TestifyConfiguration.exclusionRects] are used to exclude pixels from the comparison.
  *
- * @param configuration The configuration to use for the comparison.
+ * @param configuration - The configuration to use for the comparison.
+ * @param parallelProcessorConfiguration - The configuration for the [ParallelPixelProcessor].
  */
-internal class FuzzyCompare(private val configuration: TestifyConfiguration) {
+internal class FuzzyCompare(
+    configuration: TestifyConfiguration,
+    private val parallelProcessorConfiguration: ParallelProcessorConfiguration = ParallelProcessorConfiguration()
+) {
+    private val exclusionRects: Set<Rect> = configuration.exclusionRects
+    private val exactness: Float = configuration.exactness ?: 1f
+
+    private val analyzePixelFunction: (baselinePixel: Int, currentPixel: Int) -> Boolean =
+        if (configuration.hasExactness) { baselinePixel, currentPixel ->
+            val baselineLab = RGB.fromInt(baselinePixel).toLAB()
+            val currentLab = RGB.fromInt(currentPixel).toLAB()
+
+            val deltaE = calculateDeltaE(
+                baselineLab.l,
+                baselineLab.a,
+                baselineLab.b,
+                currentLab.l,
+                currentLab.a,
+                currentLab.b
+            )
+            ((100.0 - deltaE) / 100.0f >= exactness)
+        }
+        else { baselinePixel, currentPixel ->
+            baselinePixel == currentPixel
+        }
 
     fun compareBitmaps(baselineBitmap: Bitmap, currentBitmap: Bitmap): Boolean {
         if (baselineBitmap.height != currentBitmap.height) {
@@ -56,40 +83,14 @@ internal class FuzzyCompare(private val configuration: TestifyConfiguration) {
         }
 
         return ParallelPixelProcessor
-            .create()
+            .create(parallelProcessorConfiguration)
             .baseline(baselineBitmap)
             .current(currentBitmap)
             .analyze { baselinePixel, currentPixel, (x, y) ->
-                if (baselinePixel == currentPixel) {
-                    /* return  */ true
-                } else {
-                    var exclude = false
-                    for (rect in configuration.exclusionRects) {
-                        if (rect.contains(x, y)) {
-                            exclude = true
-                            break
-                        }
-                    }
-                    when {
-                        exclude -> true // return ^analyze
-                        configuration.hasExactness -> {
-                            val baselineLab = RGB.fromInt(baselinePixel).toLAB()
-                            val currentLab = RGB.fromInt(currentPixel).toLAB()
-
-                            val deltaE = calculateDeltaE(
-                                baselineLab.l,
-                                baselineLab.a,
-                                baselineLab.b,
-                                currentLab.l,
-                                currentLab.a,
-                                currentLab.b
-                            )
-                            ((100.0 - deltaE) / 100.0f >= configuration.exactness!!) // return ^analyze
-                        }
-
-                        else -> baselinePixel == currentPixel // return ^analyze
-                    }
-                }
+                (baselinePixel == currentPixel) || exclusionRects.any { it.contains(x, y) } || analyzePixelFunction(
+                    baselinePixel,
+                    currentPixel
+                )
             }
     }
 }