+++ Library/src/test/java/dev/testify/core/processor/ParallelPixelProcessorTest.kt
@@ -1,7 +1,7 @@
 /*
  * The MIT License (MIT)
  *
- * Copyright (c) 2023 ndtp
+ * Copyright (c) 2023-2024 ndtp
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -23,6 +23,7 @@
  */
 package dev.testify.core.processor
 
+import android.graphics.Bitmap
 import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
@@ -32,7 +33,6 @@ import kotlinx.coroutines.test.setMain
 import org.junit.After
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
-import org.junit.Before
 import org.junit.Test
 import java.util.concurrent.atomic.AtomicInteger
 
@@ -42,20 +42,22 @@ class ParallelPixelProcessorTest {
 
     private val mainThreadSurrogate = newSingleThreadContext("UI thread")
 
-    private lateinit var pixelProcessor: ParallelPixelProcessor
-
-    private fun forceSingleThreadedExecution() {
+    private fun forceSingleThreadedExecution(maxNumberOfChunkThreads: Int?): ParallelProcessorConfiguration {
         Dispatchers.setMain(mainThreadSurrogate)
-        _executorDispatcher = Dispatchers.Main
+        return ParallelProcessorConfiguration(maxNumberOfChunkThreads).apply {
+            _executorDispatcher = Dispatchers.Main
+        }
     }
 
-    @Before
-    fun setUp() {
-        forceSingleThreadedExecution()
-        pixelProcessor = ParallelPixelProcessor
-            .create()
-            .baseline(mockBitmap())
-            .current(mockBitmap())
+    private fun setUp(
+        maxNumberOfChunkThreads: Int? = null,
+        baseline: Bitmap = mockBitmap(),
+        current: Bitmap = mockBitmap()
+    ): ParallelPixelProcessor {
+        return ParallelPixelProcessor
+            .create(forceSingleThreadedExecution(maxNumberOfChunkThreads))
+            .baseline(baseline)
+            .current(current)
     }
 
     @After
@@ -65,49 +67,47 @@ class ParallelPixelProcessorTest {
     }
 
     @Test
-    fun default() {
-        maxNumberOfChunkThreads = 1
+    fun `WHEN bitmap is processed THEN analyze every pixel`() {
+        val pixelProcessor = setUp(maxNumberOfChunkThreads = 1)
 
-        val index = AtomicInteger(0)
+        val analyzed = AtomicInteger(0)
         pixelProcessor.analyze { _, _, _ ->
-            index.incrementAndGet()
+            analyzed.incrementAndGet()
             true
         }
-        assertEquals(DEFAULT_BITMAP_WIDTH * DEFAULT_BITMAP_HEIGHT, index.get())
+        assertEquals(DEFAULT_BITMAP_WIDTH * DEFAULT_BITMAP_HEIGHT, analyzed.get())
     }
 
     @Test
-    fun twoCores() {
-        maxNumberOfChunkThreads = 2
+    fun `WHEN processor has two threads THEN analyze every pixel`() {
+        val pixelProcessor = setUp(maxNumberOfChunkThreads = 2)
 
-        val index = AtomicInteger(0)
+        val analyzed = AtomicInteger(0)
         pixelProcessor.analyze { _, _, _ ->
-            index.incrementAndGet()
+            analyzed.incrementAndGet()
             true
         }
 
-        assertEquals(DEFAULT_BITMAP_WIDTH * DEFAULT_BITMAP_HEIGHT, index.get())
+        assertEquals(DEFAULT_BITMAP_WIDTH * DEFAULT_BITMAP_HEIGHT, analyzed.get())
     }
 
     @Test
-    fun oddNumberOfCores() {
-        maxNumberOfChunkThreads = 7
+    fun `WHEN there are an odd number of threads THEN analyze every pixel`() {
+        val pixelProcessor = setUp(maxNumberOfChunkThreads = 7)
 
-        val index = AtomicInteger(0)
+        val analyzed = AtomicInteger(0)
         pixelProcessor.analyze { _, _, _ ->
-            index.incrementAndGet()
+            analyzed.incrementAndGet()
             true
         }
 
-        assertEquals(DEFAULT_BITMAP_WIDTH * DEFAULT_BITMAP_HEIGHT, index.get())
+        assertEquals(DEFAULT_BITMAP_WIDTH * DEFAULT_BITMAP_HEIGHT, analyzed.get())
     }
 
     @Test
-    fun oddNumberOfPixels() {
-        maxNumberOfChunkThreads = 2
-
-        pixelProcessor = ParallelPixelProcessor
-            .create()
+    fun `WHEN there are an odd number of pixels THEN analyze every pixel`() {
+        val pixelProcessor = ParallelPixelProcessor
+            .create(ParallelProcessorConfiguration(requestedNumberOfChunkThreads = 2))
             .baseline(mockBitmap(3, 3))
             .current(mockBitmap(3, 3))
 
@@ -117,27 +117,43 @@ class ParallelPixelProcessorTest {
             0 to 2, 1 to 2, 2 to 2
         )
 
-        val index = AtomicInteger(0)
+        val analyzed = AtomicInteger(0)
         pixelProcessor.analyze { _, _, (x, y) ->
             assertTrue(expected.remove(x to y))
-            index.incrementAndGet()
+            analyzed.incrementAndGet()
             true
         }
-        assertEquals(9, index.get())
+        assertEquals(9, analyzed.get())
         assertTrue(expected.isEmpty())
     }
 
-    private fun assertPosition(index: Int, position: Pair<Int, Int>) {
-        val (x, y) = pixelProcessor.getPosition(index, DEFAULT_BITMAP_WIDTH)
+    /**
+     * Assert that the position at [index] matches the expected [position]
+     */
+    private fun ParallelPixelProcessor.assertPosition(index: Int, position: Pair<Int, Int>) {
+        val (x, y) = this.getPosition(index, DEFAULT_BITMAP_WIDTH)
         assertEquals(position, x to y)
     }
 
     @Test
-    fun multicoreChunks() {
-        maxNumberOfChunkThreads = 2
-        assertPosition(7, 7 to 0)
-        assertPosition(500, 500 to 0)
-        assertPosition(1500, 420 to 1)
-        assertPosition(2200, 40 to 2)
+    fun `WHEN using multiple threads THEN the positions map correctly`() {
+        setUp(maxNumberOfChunkThreads = 2).run {
+            assertPosition(7, 7 to 0)
+            assertPosition(500, 500 to 0)
+            assertPosition(1500, 420 to 1)
+            assertPosition(2200, 40 to 2)
+        }
+    }
+
+    @Test
+    fun `WHEN a single pixel is different THEN fail early AND do not analyze every pixel`() {
+        val pixelProcessor = setUp(maxNumberOfChunkThreads = 1)
+
+        val analyzed = AtomicInteger(0)
+        pixelProcessor.analyze { _, _, (x, y) ->
+            analyzed.incrementAndGet()
+            (x != 1) || (y != 0)
+        }
+        assertEquals(2, analyzed.get())
     }
 }